menambahkan fungsi kunci data sebelum login (tpi masih salah, nanti saya perbaiki setelah commitan ini)

+++ hapus_pengguna.php
@@ -1,4 +1,13 @@
 <?php
+session_start();
+if (!isset($_SESSION['pengguna'])) {
+    header("Location: login.php");
+    exit;
+}
+?>
+
+
+<?php
     include "koneksi.php";
     $var_id = $_GET['id'];
     $sql    = "delete from tb_pengguna where id_pengguna = '$var_id'";
